Add from_or_boxed constructor
This gives us a way to support any size future with a small value
optimization. The `from_or_boxed` constructor attempts to store the
wrapped future in `StackFuture`'s reserved space, but if it cannot then
it store the future in a box and wrap the box instead.

Note that doing this requires adding an "alloc" feature so this crate
can continue to be used in pure no_std environments.

This also changes the panic behavior of `try_from` to not panic on
alignment errors. To make it easier to determine why something went
wrong, we make the `has_*_for*` functions public. We also adjust the
panic messages in `StackFuture::from` to make it clear which conditions
failed.

Finally, we introduce a testing matrix for GitHub actions to make sure
we run tests both with and without allocation.